beta 0.2.0
All logic for patterns of tiles  that I am aware of has been added.
There may be extremely unlikely fringe scenarios that have the capability of causing the board resetter to infinitely loop; with the next commit, I will be fixing that.
I will likely, if time permits, create a GUI to wrap around this backend code, and create new code to accompany that GUI (if needed) sometime during the next quarter. The minesweeper application will be considered out of beta once that is finished and confirmed to be operational.
As the backend is fully functional now, it is considered to have exited alpha stage.